Netherite and PBR
- Updated Netherite armor
- Not quite finished this release yet

So the last release I had introduced PBR materials into the resource pack with out affecting the rest of the pack. However, I wanted to introduce emissive textures into the netherite armor.

After doing research, I found a way to do it, but it means I will need to make a optifine version, and a vanilla version.

Going to do more research to see if there is a better way to do it without needing to make 2 versions.